Split out NVSE from Installer Moved FOMOD constants and logging into FOMOD
